packaged-agent/pyproject.toml
[project]
name = "packaged-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hamza", email = "hamza.ittefaq@ssasoft.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
]

[project.scripts]
packaged-agent = "packaged_agent:main"
agents-packaged = "packaged_agent.main:run_agent"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
